move get_cos_summary from SQLQuery to cos; add AttributeError when logon() fails; other minor fixs
